requirements.txt
altgraph==0.17.2
asgiref==3.4.1
auto-py-to-exe==2.15.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
djangorestframework==3.13.1
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyJWT==2.3.0
PyMySQL==1.0.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
twilio==7.5.1
uritemplate==4.1.1
urllib3==1.26.8
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
